+++ src/main/kotlin/com/github/cnrture/quickprojectwizard/gradle/Plugin.kt
@@ -10,4 +10,5 @@ sealed class Plugin(val name: String, val id: String, val verRef: String) {
     data object Detekt : Plugin("detekt-plugin", "io.gitlab.arturbosch.detekt", "detekt")
     data object GoogleServices : Plugin("google-services", "com.google.gms.google-services", "googleServices")
     data object NavigationSafeArgs : Plugin("navigation-safe-args", "androidx.navigation.safeargs.kotlin", "navigation")
+    data object KotlinxSerialization : Plugin("kotlinx-serialization-plugin", "org.jetbrains.kotlin.plugin.serialization", "kotlin")
 }