+++ CHANGELOG.md
@@ -4,12 +4,14 @@
 
 ## [Unreleased]
 
-## [0.0.4]
+## [0.0.5]
 
 ### Added
 
-- Added Compose Multiplatform project template
-- Bugs fixed
+- Compose type safe navigation
+- Kotlin serialization
+- Target SDK 35
+- Bug fixes
 
 [Unreleased]: https://github.com/cnrture/QuickProjectWizard/compare/v0.0.4...HEAD
 [0.0.4]: https://github.com/cnrture/QuickProjectWizard/commits/v0.0.4